Move internal models one level up
