+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/internal/repositories/DefaultRoleSetupRepository.java
@@ -5,8 +5,8 @@ import org.jdbi.v3.sqlobject.statement.SqlUpdate;
 import uk.gov.hmcts.reform.amlib.enums.AccessType;
 import uk.gov.hmcts.reform.amlib.enums.RoleType;
 import uk.gov.hmcts.reform.amlib.enums.SecurityClassification;
-import uk.gov.hmcts.reform.amlib.internal.repositories.models.ResourceAttribute;
-import uk.gov.hmcts.reform.amlib.internal.repositories.models.RoleBasedAccessRecord;
+import uk.gov.hmcts.reform.amlib.internal.models.ResourceAttribute;
+import uk.gov.hmcts.reform.amlib.internal.models.RoleBasedAccessRecord;
 
 @SuppressWarnings({"LineLength", "PMD"})
 public interface DefaultRoleSetupRepository {
