+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/models/Resource.java
@@ -3,7 +3,6 @@ package uk.gov.hmcts.reform.amlib.models;
 import com.fasterxml.jackson.databind.JsonNode;
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.hmcts.reform.amlib.internal.repositories.models.ResourceDefinition;
 
 @Data
 @Builder
