+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/models/ResourceDefinition.java
@@ -1,12 +1,10 @@
-package uk.gov.hmcts.reform.amlib.internal.repositories.models;
+package uk.gov.hmcts.reform.amlib.models;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-@AllArgsConstructor
 public final class ResourceDefinition {
     private final String serviceName;
     private final String resourceType;
