+++ am-lib/src/test/java/integration/uk/gov/hmcts/reform/amlib/helpers/DatabaseHelperRepository.java
@@ -7,10 +7,10 @@ import org.jdbi.v3.sqlobject.statement.SqlUpdate;
 import uk.gov.hmcts.reform.amlib.enums.SecurityClassification;
 import uk.gov.hmcts.reform.amlib.internal.repositories.mappers.JsonPointerMapper;
 import uk.gov.hmcts.reform.amlib.internal.repositories.mappers.PermissionSetMapper;
-import uk.gov.hmcts.reform.amlib.internal.repositories.models.ResourceAttribute;
-import uk.gov.hmcts.reform.amlib.internal.repositories.models.ResourceDefinition;
-import uk.gov.hmcts.reform.amlib.internal.repositories.models.Role;
-import uk.gov.hmcts.reform.amlib.internal.repositories.models.Service;
+import uk.gov.hmcts.reform.amlib.internal.models.ResourceAttribute;
+import uk.gov.hmcts.reform.amlib.models.ResourceDefinition;
+import uk.gov.hmcts.reform.amlib.internal.models.Role;
+import uk.gov.hmcts.reform.amlib.internal.models.Service;
 
 @SuppressWarnings({"PMD", "LineLength"})
 @RegisterColumnMapper(JsonPointerMapper.class)
